refactor: using batchInsert everywhere
